requirements.txt
asgiref==3.7.2
axios==0.4.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
Django==5.0
django-cors-headers==4.3.1
djangorestframework==3.14.0
idna==3.6
joblib==1.3.2
lxml==4.9.4
markdown-it-py==3.0.0
mdurl==0.1.2
nltk==3.8.1
numpy==1.26.2
Pygments==2.17.2
pytz==2023.3.post1
regex==2023.12.25
requests==2.31.0
rich==13.7.0
scikit-learn==1.2.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
tqdm==4.66.1
tzdata==2023.3
urllib3==2.1.0
